added util functions and build fix
